delete & refactor
